+++ src/canvas/widgets/mem_basic.rs
@@ -133,7 +133,7 @@ impl Painter {
         #[cfg(feature = "gpu")]
         {
             if let Some(gpu_data) = &app_state.converted_data.gpu_data {
-                let gpu_styles = &self.colours.gpu_colour_styles;
+                let gpu_styles = &self.colours.gpu_colours;
                 let mut color_index = 0;
 
                 gpu_data.iter().for_each(|gpu_data_vec| {