+++ src/options/config/style/widget.rs
@@ -1,13 +1,13 @@
 use serde::{Deserialize, Serialize};
 
-use super::{Color, TextStyleConfig};
+use super::{ColorStr, TextStyleConfig};
 
 /// General styling for generic widgets.
 #[derive(Clone, Debug, Default, Deserialize, Serialize)]
 #[cfg_attr(feature = "generate_schema", derive(schemars::JsonSchema))]
 pub(crate) struct WidgetStyle {
-    pub(crate) border: Option<Color>,
-    pub(crate) highlighted_border_color: Option<Color>,
+    pub(crate) border: Option<ColorStr>,
+    pub(crate) highlighted_border_color: Option<ColorStr>,
     pub(crate) widget_title: Option<TextStyleConfig>,
 
     pub(crate) text: Option<TextStyleConfig>,