+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1276](https://github.com/ClementTsang/bottom/pull/1276): Add GPU process info.
 - [#1353](https://github.com/ClementTsang/bottom/pull/1353): Support selecting the average CPU graph as a default.
 - [#1430](https://github.com/ClementTsang/bottom/pull/1430): Support controlling the graph legend position for memory and network graph widgets.
+- [#1499](https://github.com/ClementTsang/bottom/pull/1499): Support bold text customization.
 
 ### Changes
 
@@ -33,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `mount_filter` is now `disk.mount_filter`.
   - `temp_filter` is now `temperature.sensor_filter`
   - `net_filter` is now `network.interface_filter`
+- [#1499](https://github.com/ClementTsang/bottom/pull/1499): Redesign how styling is configured.
 
 ### Bug Fixes